Added a base class for actions. Added several methods to bot.php: getListeners(), getListener() and getCommand()
