+++ Classes/Library/IRC/Listener/Base.php
@@ -1,20 +1,8 @@
 <?php
 namespace Library\IRC\Listener;
 
-abstract class Base
+abstract class Base extends \Library\IRC\Base
 {
-    /**
-     * Reference to the IRC Connection.
-     * @var \Library\IRC\Connection
-     */
-    protected $connection = null;
-
-    /**
-     * Reference to the IRC Bot
-     * @var \Library\IRC\Bot
-     */
-    protected $bot = null;
-
     /**
      * The listener main function.
      * @param string The raw data the listener receives.
@@ -41,22 +29,6 @@ abstract class Base
     }
 
     /**
-     * Set's the IRC Connection, so we can use it to send data to the server.
-     * @param \Library\IRC\Connection $ircConnection
-     */
-    public function setIRCConnection( \Library\IRC\Connection $ircConnection ) {
-        $this->connection = $ircConnection;
-    }
-
-    /**
-     * Set's the IRC Bot, so we can use it to send data to the server.
-     * @param \Library\IRCBot $ircBot
-     */
-    public function setIRCBot( \Library\IRC\Bot $ircBot ) {
-        $this->bot = $ircBot;
-    }
-
-    /**
      * Returns the raw data as an array of parsed data.
      * @param string $data
      * @return array
@@ -78,43 +50,5 @@ abstract class Base
         ///////////////////////////////////////////////////////////////////////////////////////////////
 
         return array_map($func, $args);
-        
-                /**
-         * Fetches data from $uri
-         *
-         * @param string $uri
-         * @return string
-         */
-    }
-    
-    /**
-     * Fetches data from $uri
-     *
-     * @param string $uri
-     * @return string
-     */
-    protected function fetch($uri) {
-
-        $this->bot->log("Fetching from URI: " . $uri);
-
-        // create curl resource
-        $ch = curl_init();
-
-        // set url
-        curl_setopt($ch, CURLOPT_URL, $uri);
-
-        //return the transfer as a string
-        curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-        curl_setopt($ch, CURLOPT_FRESH_CONNECT, 1);
-
-        // $output contains the output string
-        $output = curl_exec($ch);
-
-        // close curl resource to free up system resources
-        curl_close($ch);
-
-        $this->bot->log("Data fetched: " . $output);
-        
-        return $output;
     }
 }
